#821 [REFACTOR] shared/component/suspense 분류 완료
