+++ src/components/Carousel/CarouselErrorFallback/CarouselErrorFallback.jsx
@@ -1,4 +1,4 @@
-import { ServerErrorFallback } from '@/components';
+import { ServerErrorFallback } from '@/shared/component';
 
 export default function CarouselErrorFallback({ error, resetErrorBoundary }) {
   return <ServerErrorFallback reset={resetErrorBoundary} />;
