+++ src/components/Comment/Comments/CommentsErrorFallback.jsx
@@ -1,4 +1,4 @@
-import { ServerErrorFallback } from '@/components';
+import { ServerErrorFallback } from '@/shared/component';
 
 export default function CommentsErrorFallback({ error, resetErrorBoundary }) {
   return <ServerErrorFallback reset={resetErrorBoundary} />;
