--- src/components/ServerErrorFallback/index.js
@@ -1,1 +1,0 @@
-export { default as ServerErrorFallback } from './ServerErrorFallback';
