--- src/components/ShimmerEffect/index.js
@@ -1,1 +1,0 @@
-export { default as ShimmerEffect } from './ShimmerEffect';
