+++ src/pages/ExamReviewsPage/components/SearchExamReviews/SearchExamReviewsErrorFallback.jsx
@@ -1,5 +1,4 @@
-import { FetchLoading } from '@/shared/component';
-import { ServerErrorFallback } from '@/components';
+import { FetchLoading, ServerErrorFallback } from '@/shared/component';
 
 export default function SearchExamReviewsErrorFallback({
   error,
