+++ src/pages/MainPage/components/HomeBesookt/HomeBesooktSkeleton.jsx
@@ -1,4 +1,4 @@
-import { ShimmerEffect } from '@/components';
+import { ShimmerEffect } from '@/shared/component';
 
 import styles from './HomeBesooktSkeleton.module.css';
