+++ src/pages/MainPage/components/HomeCard/HomeCardErrorFallback.jsx
@@ -1,4 +1,4 @@
-import { ServerErrorFallback } from '@/components';
+import { ServerErrorFallback } from '@/shared/component';
 
 export default function HomeCardErrorFallback({ error, resetErrorBoundary }) {
   return <ServerErrorFallback reset={resetErrorBoundary} />;
