+++ src/pages/PostsPage/components/Posts/PostsErrorFallback.jsx
@@ -1,4 +1,4 @@
-import { ServerErrorFallback } from '@/components';
+import { ServerErrorFallback } from '@/shared/component';
 
 export default function PostsErrorFallback({ error, resetErrorBoundary }) {
   return <ServerErrorFallback reset={resetErrorBoundary} />;
