+++ src/shared/component/suspense/ShimmerEffect/ShimmerEffect.module.css
